Change int to ll

+++ Data_Structures/Lazy_Segtree.cpp
@@ -10,7 +10,7 @@ void update(ll id,ll l,ll r,ll x,ll y,ll val){
     if(x<=l && r<=y)add(id,l,r,val);
     else if(l>y || r<x)return;
     else{
-        int mid=(l+r)>>1;
+        ll mid=(l+r)>>1;
         add(id<<1,l,mid,lazy[id]);
         add(id<<1|1,mid+1,r,lazy[id]);
         lazy[id]=0;
@@ -19,11 +19,11 @@ void update(ll id,ll l,ll r,ll x,ll y,ll val){
         segt[id]=max(segt[id<<1],segt[id<<1|1]);
     }
 }
-int query(ll id,ll l,ll r,ll x,ll y){
+ll query(ll id,ll l,ll r,ll x,ll y){
     if(x<=l && r<=y)return segt[id];
     else if(l>y || r<x)return NEUTRAL;
     else{
-        int mid=(l+r)>>1;
+        ll mid=(l+r)>>1;
         add(id<<1,l,mid,lazy[id]);
         add(id<<1|1,mid+1,r,lazy[id]);
         lazy[id]=0;
